backend/requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
python-multipart==0.0.9
Pillow==9.4.0
tensorflow==2.12.0
numpy>=1.22,<1.24
